@@ -4,8 +4,8 @@
   "private": true,
   "dependencies": {
     "node-sass": "^4.12.0",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
     "react-modal": "^3.1.10",
     "react-router": "^4.2.0",
     "react-router-dom": "^4.2.2",
